Brace yourself, major refactor is coming!
Well, it already come. This closes #53.
