+++ src/client/components/developer-list/developer-card/DeveloperCard.js
