+++ src/client/pages/developers/Developers.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Route, Redirect, NavLink } from 'react-router-dom'
-import Popular from './popular/Popular'
-import RecentlyJoined from './recently-joined/RecentlyJoined'
-import SubNavLink from '../sub-nav-link/SubNavLink'
-import store from '../../store/store'
+import Popular from '~/pages/popular-developers/PopularDevelopers'
+import RecentlyJoined from '~/pages/recently-joined-developers/RecentlyJoinedDevelopers'
+import SubNavLink from '~/components/sub-nav-link/SubNavLink'
+import store from '~/store/store'
 
 const Developers = (props) => (
   <div id="developers">
